requirements.txt
Django==2.1
Unidecode==1.0.23
pymongo-3.8.0

